+++ viz/rendering-app/src/androidTest/java/io/data2viz/viz/RenderingTestSupport.kt
@@ -348,7 +348,7 @@ val allRenderingTests = listOf(
                 y = 200.0
                 radius = 100.0
                 fill = null
-                stroke = Colors.Web.red
+                strokeColor = Colors.Web.red
                 strokeWidth = 20.0
             }
         },
@@ -358,7 +358,7 @@ val allRenderingTests = listOf(
                 y = 200.0
                 radius = 100.0
                 fill = null
-                stroke = Colors.Web.red
+                strokeColor = Colors.Web.red
             }
         },
         renderingTest("circle4") {
@@ -367,7 +367,7 @@ val allRenderingTests = listOf(
                 y = 200.0
                 radius = 100.0
                 fill = 0xfdc658.col
-                stroke = 0x0c0887.col
+                strokeColor = 0x0c0887.col
                         .withAlpha(50.pct)
                 strokeWidth = 40.0
             }
@@ -386,7 +386,7 @@ val allRenderingTests = listOf(
                 arc(pos.first, pos.second, 25.0, .0, it * (2 * PI / 8.0), false)
                 closePath()
                 fill = Colors.Web.grey
-                stroke = null
+                strokeColor = null
             }
         }
     },
@@ -401,7 +401,7 @@ val allRenderingTests = listOf(
                 arc(pos.first, pos.second, 25.0, .0, -it * (2 * PI / 8.0), false)
                 closePath()
                 fill = Colors.Web.grey
-                stroke = null
+                strokeColor = null
             }
         }
     },
@@ -416,7 +416,7 @@ val allRenderingTests = listOf(
                 arc(pos.first, pos.second, 25.0, .0, it * (2 * PI / 8.0), true)
                 closePath()
                 fill = Colors.Web.grey
-                stroke = null
+                strokeColor = null
             }
         }
     },
@@ -431,7 +431,7 @@ val allRenderingTests = listOf(
                 arc(pos.first, pos.second, 25.0, .0, -it * (2 * PI / 8.0), true)
                 closePath()
                 fill = Colors.Web.grey
-                stroke = null
+                strokeColor = null
             }
         }
     },
@@ -448,7 +448,7 @@ val allRenderingTests = listOf(
                 arc(pos.first, pos.second, 25.0, it * posNeg / 10.0, -posNeg * it * (2 * PI / 8.0), false)
                 closePath()
                 fill = Colors.Web.grey
-                stroke = null
+                strokeColor = null
             }
         }
     },
@@ -465,7 +465,7 @@ val allRenderingTests = listOf(
                 arc(pos.first, pos.second, 25.0, it * posNeg / 10.0, -posNeg * it * (2 * PI / 8.0), true)
                 closePath()
                 fill = Colors.Web.grey
-                stroke = null
+                strokeColor = null
             }
         }
     },
@@ -484,7 +484,7 @@ val allRenderingTests = listOf(
                 lineTo(pos.first + 15.0, pos.second + 15.0)
                 closePath()
                 fill = Colors.Web.grey
-                stroke = Colors.Web.blue
+                strokeColor = Colors.Web.blue
             }
         }
     },
@@ -503,7 +503,7 @@ val allRenderingTests = listOf(
                 lineTo(pos.first + 15.0, pos.second + 15.0)
                 closePath()
                 fill = Colors.Web.grey
-                stroke = Colors.Web.blue
+                strokeColor = Colors.Web.blue
             }
         }
     },
@@ -530,7 +530,7 @@ val allRenderingTests = listOf(
             lineTo(300.0, 312.0)
             closePath()
             fill = Colors.Web.grey
-            stroke = Colors.Web.blue
+            strokeColor = Colors.Web.blue
         }
     },
 
@@ -557,7 +557,7 @@ val allRenderingTests = listOf(
             lineTo(300.0, 312.0)
             closePath()
             fill = colors.grey
-            stroke = colors.blue
+            strokeColor = colors.blue
         }
     },*/
 
@@ -568,7 +568,7 @@ val allRenderingTests = listOf(
                 lineTo(60.0, 20.0)
                 moveTo(80.0, 40.0)
                 lineTo(100.0, 20.0)
-                stroke = Colors.Web.red
+                strokeColor = Colors.Web.red
             }
         },
         renderingTest("path.rect") {