+++ viz/src/jsMain/kotlin/io/data2viz/viz/JsCircleRenderer.kt
@@ -29,7 +29,7 @@ public fun CircleNode.render(context: CanvasRenderingContext2D) {
         context.fill()
     }
 
-    stroke?.let {
+    strokeColor?.let {
         context.stroke()
     }