+++ viz/src/main/java/io/data2viz/viz/AndroidLineRenderer.kt
@@ -20,7 +20,7 @@ package io.data2viz.viz
 import android.graphics.*
 
 fun LineNode.render(renderer: AndroidCanvasRenderer) {
-	stroke?.let {
+	strokeColor?.let {
 		paint.style = Paint.Style.STROKE
 		paint.strokeWidth = strokeWidth?.toFloat() ?: 1f
 		it.updatePaint(paint, renderer)