+++ viz/src/main/java/io/data2viz/viz/AndroidPathRenderer.kt
@@ -118,7 +118,7 @@ fun PathNode.render(renderer: AndroidCanvasRenderer) {
         canvas.drawPath(path, paint)
     }
 
-    stroke?.let {
+    strokeColor?.let {
         paint.style = Paint.Style.STROKE
         it.updatePaint(paint, renderer)
         canvas.drawPath(path, paint)