+++ viz/src/main/java/io/data2viz/viz/AndroidTextRenderer.kt
@@ -38,7 +38,7 @@ internal fun TextNode.render(renderer: AndroidCanvasRenderer) {
 				y.dp - dy,
 				paint )
 		}
-		stroke?.let {
+		strokeColor?.let {
 			paint.style = Paint.Style.STROKE
 			it.updatePaint(paint, renderer)
 			canvas.drawText(